+++ tools/carbon/modules.h
@@ -27,5 +27,7 @@ bool moduleJs2CabInvoke(int argc, char **argv, FILE *file, carbon_cmdopt_mgr_t *
 bool moduleViewCabInvoke(int argc, char **argv, FILE *file, carbon_cmdopt_mgr_t *manager);
 bool moduleInspectInvoke(int argc, char **argv, FILE *file, carbon_cmdopt_mgr_t *manager);
 bool moduleCab2JsInvoke(int argc, char **argv, FILE *file, carbon_cmdopt_mgr_t *manager);
+bool moduleListInvoke(int argc, char **argv, FILE *file, carbon_cmdopt_mgr_t *manager);
+
 
 #endif